kafka/kafka-python/requirements.txt
mysql-connector-python
pandas==1.2.2
kafka-python